% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_diff.R
\name{angle_diff}
\alias{angle_diff}
\title{Angle Difference}
\usage{
angle_diff(theta)
}
\arguments{
\item{theta}{Vector of angles of size n.}
}
\value{
A vector of size `n` whose first entry is NA, the second is the
  difference between the 2nd and 1st, and so on.
}
\description{
Compute the angle difference. The Output are values between 0 and 180
degrees.
}
\examples{
angle_diff(c(330, 15, 60, 15))
}
