% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_summaries.R
\name{calculate_longest_na_seq}
\alias{calculate_longest_na_seq}
\title{Size of longest sequence of NAs}
\usage{
calculate_longest_na_seq(x)
}
\arguments{
\item{x}{Vector with NAs.}
}
\value{
Integer with the size of the longest consecutive NAs in vector x.
}
\description{
Calculate the longest sequence of NAs in a vector. Zero si returned if there are no NAs.
}
\examples{
## Max = 3
## NA sequences: 4
x <- c(1, NA, 2, NA, NA, 3, 4, NA, NA, NA, 5, 6, 7, NA)
calculate_longest_na_seq(x)
calculate_number_of_na_seqs(x)
}
