% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_summaries.R
\name{calculate_number_of_na_seqs}
\alias{calculate_number_of_na_seqs}
\title{Total NA sequences}
\usage{
calculate_number_of_na_seqs(x)
}
\arguments{
\item{x}{Vector with, potentially, missing values.}
}
\value{
An integer with the total number of sequences.
}
\description{
Compute the number of sequences of NAs. Returns zero if there are no missing values.
}
\examples{
## Max = 3
## NA sequences: 4
x <- c(1, NA, 2, NA, NA, 3, 4, NA, NA, NA, 5, 6, 7, NA)
calculate_longest_na_seq(x)
calculate_number_of_na_seqs(x)
}
