% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_turn.R
\name{classify_turn}
\alias{classify_turn}
\title{Classify turns}
\usage{
classify_turn(crossprod_sign, heading_difference, angle_threshold)
}
\arguments{
\item{crossprod_sign}{a}

\item{heading_difference}{b}

\item{angle_threshold}{c}
}
\value{
we
}
\description{
Classify turns
}
\examples{
1 + 1
}
