% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gps_heading.R
\name{compute_gps_heading}
\alias{compute_gps_heading}
\title{Compute GPS Heading}
\usage{
compute_gps_heading(latitude, longitude)
}
\arguments{
\item{latitude}{Vector of latitudes of size n.}

\item{longitude}{Vector of longitudes of size n.}
}
\value{
Vector of headings (between 0 and 360) of size n. The last entry is \code{NA}.
}
\description{
Calculate the heading of two vectors of coordinates.
}
\examples{
cities <-
tibble::tribble(
  ~city, ~latitude, ~longitude,
  "Las Vegas", 36.1699, -115.1398,
  "San Diego", 32.7157, -117.1611,
  "Los Angeles", 34.0522, -118.2437,
  "San Francisco", 37.7749, -122.4194,
  "Portland", 45.5122, -122.6587,
  "Seattle", 47.6062, -122.3321,
  "Vancouver", 49.2827, -123.1207,
  "Anchorage", 61.2181, -149.9003
)

with(cities, compute_gps_heading(latitude, longitude))
}
