% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_prod_sign.R
\name{cross_prod_sign}
\alias{cross_prod_sign}
\title{Cross Product Sign}
\usage{
cross_prod_sign(lat, lon)
}
\arguments{
\item{lat}{Vector of latitudes of size n.}

\item{lon}{Vector of longitudes of size n.}
}
\value{
Vector of cross product signs of size n. The first element and last
  elements are NA.
}
\description{
Calculate the sign of the cross product.
}
\details{
Negative values are associated with right turns while positive values with
left turns.
}
\examples{
cities <-
tibble::tribble(
  ~city, ~latitude, ~longitude,
  "Las Vegas", 36.1699, -115.1398,
  "San Diego", 32.7157, -117.1611,
  "Los Angeles", 34.0522, -118.2437,
  "San Francisco", 37.7749, -122.4194,
  "Portland", 45.5122, -122.6587,
  "Seattle", 47.6062, -122.3321,
  "Vancouver", 49.2827, -123.1207,
  "Anchorage", 61.2181, -149.9003
)

with(cities, cross_prod_sign(latitude, longitude))
}
